+++ brixterm/constants.py
@@ -3,7 +3,7 @@ DEFAULT_GPT_MODEL = "gpt-4o-mini"
 CMD_HIST_SIZE = 10
 CHAT_HIST_SIZE = 10
 TERM_INPUT_PREFIX = (
-    "\033[1;32m🤖BxT:\033[0m"  # venv: bright green
+    "\033[1;32m🤖 BxT:\033[0m"  # venv: bright green
     "\033[1;36m{}\033[0m "  # user@host: cyan
     "\033[1;34m{}\033[0m"  # cwd: blue
     "> "  # prompt symbol