+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+[2025/07/27]
+- light mode fixed for mac terminal
+- aaa mode added for chat with clipboard copied context
+
 [2025/07/26]
 - initial implementation
 - terminal with special commands